WIP: Add: Reactions
It works for newly received reactions, but after initial sync,
reactions that happened in the past are not displayed.  I think it's
because the related events aren't found in the room's timeline, but I
tried to fix that, and it still doesn't work.

I'm guessing there are some assumptions that I'm making wrongly, or
something that I don't understand about how the server sends events.
We may have to save a list of certain types of events and process them
after all other events have been processed.  Ugh.

The good news is that EWOC makes it pretty easy and reliable to update
messages in the buffer.