WIP: Call -room---process-events in -room--buffer
This almost seems to work, in that reactions from old timeline
events are displayed when the buffer is made...or not?  It seems to
work in some cases, but not in others, like when retro-loading...

The big issue now is that the reaction events cause blank events to be
inserted into the buffer.  Fixing that will require conditionally
inserting events, which probably means moving message event handling
into the defevent macro, which will require some more refactoring...