WIP: Fix: (ement-room--user-display-name) Cache other name types
WIP: This may currently cause some displaynames to not be shown
correctly when loading older messages, so not merging yet.  This gets
complicated...

When a room has many events (e.g. over 1,000), searching the events
for "displayname" events can take a long time.  If we only cache
displaynames from such events, then users without an explicitly set
displayname will cause this search to happen every time this function
is called, e.g. every time a message from such a user is formatted.

So rather than only caching a name from a "displayname" event, we
cache whatever we calculate, and we clear that cache whenever a new
"displayname" event is handled